* Server add jaeger tracing

+++ Server/clean.sh
@@ -7,5 +7,10 @@ rm -f *.log
 rm -rf ./build
 rm -rf ./data
 rm -rf ./storage/data
-rm -rf /var/tmp/n3-privacy/meta/
-rm -rf /var/tmp/n3-privacy/badger/
+
+if [ $# -gt 0 ]; then
+  if [ $1 = "rmdb" ]; then
+    rm -rf /var/tmp/n3-privacy/meta/
+    rm -rf /var/tmp/n3-privacy/badger/
+  fi
+fi
